+++ ovp_core/serializers/cause.py
@@ -1,6 +1,13 @@
 from ovp_core import models
 from ovp_core import validators
 from rest_framework import serializers
+from ovp_uploads.serializers import UploadedImageSerializer
+
+class FullCauseSerializer(serializers.ModelSerializer):
+  image = UploadedImageSerializer()
+  class Meta:
+    fields = ['id', 'name', 'image']
+    model = models.Cause
 
 class CauseSerializer(serializers.ModelSerializer):
 	class Meta: